Fix hydration by moving code into server-side file
