+++ src/components/ContributorList.astro
@@ -1,5 +1,5 @@
 ---
-import { cachedFetch } from '../util';
+import { cachedFetch } from '../util-server';
 import UIString from './UIString.astro';
 
 export interface Props {
