+++ src/components/Since.astro
@@ -1,7 +1,7 @@
 ---
 import UIString from './UIString.astro';
 import Badge from './Badge.astro';
-import { cachedFetch } from '../util';
+import { cachedFetch } from '../util-server';
 
 export interface Props {
   v: string;
