+++ src/components/Version.astro
@@ -1,5 +1,5 @@
 ---
-import { cachedFetch } from "../util";
+import { cachedFetch } from '../util-server';
 
 export interface Props {
 	pkgName: string;
