+++ src/util.ts
@@ -1,5 +1,3 @@
-import EleventyFetch from '@11ty/eleventy-fetch';
-
 export function getLanguageFromURL(pathname: string) {
 	const langCodeMatch = pathname.match(/\/([a-z]{2}-?[a-z]{0,2})\//);
 	return langCodeMatch ? langCodeMatch[1] : 'en';
@@ -41,42 +39,4 @@ export function unescapeHtml(escapedString: string) {
 			.replace(/&gt;/g, ">")
 			.replace(/&quot;/g, "\"")
 			.replace(/&#039;/g, "'");
-}
-
-export type CachedFetchOptions = {
-	duration?: string;
-	verbose?: boolean;
-};
-
-export async function cachedFetch(
-	url: string,
-	fetchOptions = {},
-	{ duration = '5m', verbose = false }: CachedFetchOptions = {}
-) {
-	let status = 200;
-	let statusText = 'OK';
-	let buffer: Buffer | undefined;
-
-	try {
-		buffer = await EleventyFetch(url, {
-			duration,
-			verbose,
-			type: 'buffer',
-			fetchOptions,
-		});
-	} catch (error) {
-		const msg: string = error?.message || error.toString();
-		const matches = msg.match(/^Bad response for (.*) \(.*?\): (.*)$/);
-		status = parseInt(matches?.[2] || '') || 404;
-		statusText = matches?.[3] || msg;
-	}
-
-	return {
-		ok: status >= 200 && status <= 299,
-		status,
-		statusText,
-		body: buffer,
-		json: () => buffer && JSON.parse(buffer.toString()),
-		text: () => buffer?.toString(),
-	};
 }
